move js files into folder
